+++ simulator_rs/src/card/stats.rs
@@ -6,17 +6,17 @@ pub struct Stats {
     pub pure: u32,
     pub cool: u32,
     #[serde(default = "default_sis_base")]
-    pub sis_base: u32,
+    pub sis_base: u8,
     #[serde(default = "default_sis_max")]
-    pub sis_max: u32,
+    pub sis_max: u8,
     pub image: Option<String>,
 }
 
-fn default_sis_base() -> u32 {
+fn default_sis_base() -> u8 {
     1
 }
 
-fn default_sis_max() -> u32 {
+fn default_sis_max() -> u8 {
     1
 }
